devide to map
